% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{combine_training_and_eval_data}
\alias{combine_training_and_eval_data}
\title{Combine training and evaluation data for
postprocessing.}
\usage{
combine_training_and_eval_data(train_dat, eval_dat)
}
\arguments{
\item{train_dat}{Training data, as a \code{\link[tibble:tibble]{tibble}}.}

\item{eval_dat}{Evaluation data, as a \code{\link[tibble:tibble]{tibble}}.}

\item{disease_name}{Name of the target disease.
One of \code{"COVID-19"} or \code{"Influenza"}.}
}
\value{
The combined data, as a \code{\link[tibble:tibble]{tibble}}.
}
\description{
Combine training and evaluation data for
postprocessing.
}
