% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_utils.R
\docType{data}
\name{disease_map_lower}
\alias{disease_map_lower}
\title{Utilities for handling and parsing directory names
based on pyrenew-hew pipeline conventions.}
\format{
An object of class \code{character} of length 3.
}
\usage{
disease_map_lower
}
\description{
Utilities for handling and parsing directory names
based on pyrenew-hew pipeline conventions.
}
\keyword{datasets}
