% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_loc_forecast.R
\name{epiweekly_samples_from_daily}
\alias{epiweekly_samples_from_daily}
\title{Aggregate daily samples to epiweekly samples}
\usage{
epiweekly_samples_from_daily(
  daily_samples,
  variables_to_aggregate = "observed_ed_visits",
  required_columns
)
}
\arguments{
\item{daily_samples}{Tibble of daily samples}

\item{variables_to_aggregate}{Name of variable to aggregate.
Defaults to "observed_ed_visits"}

\item{required_columns}{Character vector of required columns}
}
\value{
Tibble of aggregated epiweekly samples
}
\description{
Filters daily samples for specified variables,
aggregates them to epiweekly resolution,
and returns a tidy table with required columns.
}
