% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_submission.R
\name{find_multiple_options}
\alias{find_multiple_options}
\title{Find multiple forecast options for a given target and location
in a hubverse table.}
\usage{
find_multiple_options(df, model_run_id_columns)
}
\arguments{
\item{df}{Hubverse table to examine.}

\item{model_id_columns}{Columns that serve as a primary
key for individual model runs.}
}
\description{
Find multiple forecast options for a given target and location
in a hubverse table.
}
