% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_utils.R
\name{get_all_forecast_dirs}
\alias{get_all_forecast_dirs}
\title{Get forecast directories.}
\usage{
get_all_forecast_dirs(dir_of_forecast_dirs, diseases)
}
\arguments{
\item{dir_of_forecast_dirs}{Directory in which to look for
subdirectories representing individual forecast date / pathogen /
dataset combinations.}

\item{diseases}{Names of the diseases to match, as a vector of strings,
or a single disease as a string.}
}
\value{
A vector of paths to the forecast subdirectories.
}
\description{
Get all the subdirectories within a parent directory
that match the pattern for a forecast run for a
given disease and optionally a given report date.
}
