% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_utils.R
\name{get_all_model_batch_dirs}
\alias{get_all_model_batch_dirs}
\title{Get forecast directories.}
\usage{
get_all_model_batch_dirs(dir_of_batch_dirs, diseases)
}
\arguments{
\item{dir_of_batch_dirs}{Directory in which to look for
"model batch" directories, each of which represents an
individual forecast date / pathogen / dataset combination.}

\item{diseases}{Names of the diseases to match, as a vector of strings,
or a single disease as a string.}
}
\value{
A vector of paths to the forecast subdirectories.
}
\description{
Get all the subdirectories within a parent directory
that match the pattern for a forecast run for a
given disease and optionally a given report date.
}
