% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{group_lab_site_index_to_name}
\alias{group_lab_site_index_to_name}
\title{Convert lab_site_index to lab_site_name}
\usage{
group_lab_site_index_to_name(lab_site_index, lab_site_index_to_name_map)
}
\arguments{
\item{lab_site_index}{integer vector of lab site indices}

\item{lab_site_index_to_name_map}{a tibble mapping lab_site_index to names}
}
\value{
a vector of lab site names
}
\description{
Convert lab_site_index to lab_site_name
}
\examples{
group_lab_site_index_to_name_map(
  7,
  lab_site_index_to_name_map
)
}
