% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{group_time_index_to_date}
\alias{group_time_index_to_date}
\title{Convert group time index to date}
\usage{
group_time_index_to_date(
  group_time_index,
  variable,
  first_nssp_date,
  first_nhsn_date,
  first_nwss_date,
  nhsn_step_size
)
}
\arguments{
\item{group_time_index}{integer vector of group time indices}

\item{variable}{variable name}

\item{first_nssp_date}{first date in the nssp training data}

\item{first_nhsn_date}{first date in the nhsn training data}

\item{first_nwss_date}{first date in the nwss training data}

\item{nhsn_step_size}{step size for nhsn data}
}
\value{
a vector of dates
}
\description{
Convert group time index to date
}
\examples{
group_time_index_to_date(
  3, "observed_hospital_admissions",
  "2024-01-01", "2024-01-01", "2024-01-01", 7
)
}
