% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_utils.R
\name{load_training_data}
\alias{load_training_data}
\title{Utilities for timeseries forecasting scripts in the cfa-stf-routine-forecasting
pipeline.
Load Training Data for Timeseries Forecasting}
\usage{
load_training_data(model_dir, data_name = "combined_data", epiweekly = FALSE)
}
\arguments{
\item{model_dir}{Path to the
directory containing model run data.}

\item{epiweekly}{Logical. Indicate epiweekly (TRUE) or daily (FALSE) data.
Default \code{FALSE}.}

\item{base_data_name}{Base name of the data file to load.}
}
\value{
A list with:
\code{data} (processed training data tibble),
\code{geo_value},
\code{disease},
\code{resolution} ("daily" or "epiweekly"), and
\code{prefix} (file prefix based on resolution).
}
\description{
Loads and preprocesses training data for timeseries
forecasting models in the format expected by
the forecasting functions.
}
