% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast_figure.R
\name{make_forecast_figure}
\alias{make_forecast_figure}
\title{Make Forecast Figure}
\usage{
make_forecast_figure(
  target_variable,
  combined_dat,
  forecast_ci,
  data_vintage_date,
  y_transform = "identity",
  display_cutpoints = TRUE,
  highlight_dates = NULL,
  highlight_labels = NULL
)
}
\arguments{
\item{target_variable}{a variable matching the .variable columns
in \code{combined_dat} and \code{forecast_ci}}

\item{combined_dat}{\code{combined_dat} from the result of
\code{\link[=process_state_forecast]{process_state_forecast()}}}

\item{forecast_ci}{\code{forecast_ci} from the result of
\code{\link[=process_state_forecast]{process_state_forecast()}}}

\item{data_vintage_date}{date that the data was collected}

\item{y_transform}{a character passed as the transform argument to
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}

\item{display_cutpoints}{a logical indicating whether to include cutpoints
relevant to the \code{target_variable}.}

\item{highlight_dates}{a vector of dates to highlight on the plot}

\item{highlight_labels}{a vector of labels to display at the
\code{highlight_dates}}
}
\value{
a ggplot object
}
\description{
Make Forecast Figure
}
