% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dir_name.R
\name{parse_model_batch_dir_name}
\alias{parse_model_batch_dir_name}
\title{Parse the name of a model batch directory
(i.e. a directory representing a single
report date and disease pair, but potentially
with fits for multiple locations), returning
a named list of quantities of interest.}
\usage{
parse_model_batch_dir_name(model_batch_dir_name)
}
\arguments{
\item{model_batch_dir_name}{Name of the model batch
directory (not the full path to it, just the directory
base name) to parse.}
}
\value{
A list of quantities: \code{disease}, \code{report_date},
\code{first_training_date}, and \code{last_training_date}.
}
\description{
Parse the name of a model batch directory
(i.e. a directory representing a single
report date and disease pair, but potentially
with fits for multiple locations), returning
a named list of quantities of interest.
}
