% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dir_name.R
\name{parse_model_run_dir_path}
\alias{parse_model_run_dir_path}
\title{Parse path to a model run directory
(i.e. a directory representing a run for a
particular location, disease, and reference
date, and extract key quantities of interest.}
\usage{
parse_model_run_dir_path(model_run_dir_path)
}
\arguments{
\item{model_run_dir_path}{Path to parse.}
}
\value{
A list of parsed attributes:
\code{location}, \code{disease}, \code{report_date},
\code{first_training_date}, and \code{last_training_date}.
}
\description{
Parse path to a model run directory
(i.e. a directory representing a run for a
particular location, disease, and reference
date, and extract key quantities of interest.
}
