% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_utils.R
\name{parse_pyrenew_model_name}
\alias{parse_pyrenew_model_name}
\title{Parse PyRenew Model Name}
\usage{
parse_pyrenew_model_name(pyrenew_model_name)
}
\arguments{
\item{pyrenew_model_name}{name of a pyrenew model ("pyrenew_h", "pyrenew_he",
"pyrnew_hew", etc)}
}
\value{
a named logical vector indicating which components are present
}
\description{
Parse PyRenew Model Name
}
\examples{
parse_pyrenew_model_name("pyrenew_h")
}
