% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_utils.R
\name{parse_variable_name}
\alias{parse_variable_name}
\title{Parse variable name.}
\usage{
parse_variable_name(variable_name)
}
\arguments{
\item{variable_name}{Character. Name of the variable to parse.}
}
\value{
A list containing:
\itemize{
\item \code{proportion}: Logical. Indicates if the variable represents a proportion.
\item \code{core_name}: Character. A simplified name for the variable.
\item \code{full_name}: Character. A formatted name for the variable.
\item \code{y_axis_labels}: Function. A suitable label function for axis formatting.
}
}
\description{
Convert a variable name into a descriptive label for display in plots.
}
\examples{
parse_variable_name("prop_hospital_admissions")
}
