% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_utils.R
\name{path_up_to}
\alias{path_up_to}
\title{Get path up to a specific directory.}
\usage{
path_up_to(path, up_to)
}
\arguments{
\item{path}{A character vector of paths.}

\item{up_to}{A character string specifying the directory name}
}
\value{
A character vector of paths that go up to the specified directory.
}
\description{
Get path up to a specific directory.
}
