% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{pivot_ed_visit_df_longer}
\alias{pivot_ed_visit_df_longer}
\title{Pivot a data table of counts and proportions of
ED visits to long format.}
\usage{
pivot_ed_visit_df_longer(df)
}
\arguments{
\item{df}{data frame to pivot. Should have columns
\code{"Disease"}, \code{"Other"}, and \code{"prop_disease_ed_visits"}.}
}
\value{
the pivoted data frame, with disease names in
a column named \code{disease} and counts / proportions in
a column named \code{.value}.
}
\description{
Pivot a data table of counts and proportions of
ED visits to long format.
}
