% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_hubverse_tbl.R
\name{prelim_samples_to_hubverse}
\alias{prelim_samples_to_hubverse}
\title{Convert preliminary samples to Hub format}
\usage{
prelim_samples_to_hubverse(
  prelim_samples,
  output_type = c("quantiles", "samples", "both")
)
}
\arguments{
\item{prelim_samples}{A data frame of preliminary samples, in the format output by \code{raw_samples_to_prelim()}}

\item{output_type}{The type of output to produce. One of "quantiles", "samples", or "both". If "quantiles", produces quantiles in the format required by the Hub. If "samples", produces samples in the format required by the Hub. If "both", produces both quantiles and samples.}
}
\value{
A data frame in the format required by the Hub, either as quantiles, samples, or both, depending on the \code{output_type} parameter.
}
\description{
Convert preliminary samples to Hub format
}
