% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_hubverse_tbl.R
\name{prelim_to_hub_quantiles}
\alias{prelim_to_hub_quantiles}
\title{Convert preliminary samples to Hub format as quantiles}
\usage{
prelim_to_hub_quantiles(prelim_samples)
}
\arguments{
\item{prelim_samples}{A data frame of preliminary samples, in the format output by \code{raw_samples_to_prelim()}}
}
\value{
A data frame in the format required by the Hub, with quantiles as the output type.
}
\description{
Convert preliminary samples to Hub format as quantiles
}
