% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_loc_forecast.R
\name{process_model_samples.pyrenew}
\alias{process_model_samples.pyrenew}
\title{Process PyRenew model samples}
\usage{
\method{process_model_samples}{pyrenew}(
  model_type,
  model_run_dir,
  model_name,
  ts_samples = NULL,
  required_columns_e,
  n_forecast_days,
  ...
)
}
\arguments{
\item{model_type}{Character string indicating model type}

\item{model_run_dir}{Model run directory}

\item{model_name}{Name of directory containing model outputs}

\item{ts_samples}{Timeseries samples (if available)}

\item{required_columns_e}{Required columns for output}

\item{n_forecast_days}{Number of forecast days}

\item{...}{Additional arguments (unused)}
}
\value{
Tibble of PyRenew model samples
}
\description{
Process PyRenew model samples
}
