% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{process_state_forecast}
\alias{process_state_forecast}
\title{Process state forecast}
\usage{
process_state_forecast(model_run_dir, save = TRUE)
}
\arguments{
\item{model_run_dir}{Model run directory}

\item{save}{Logical indicating whether or not to save}
}
\value{
a list with three tibbles: combined_dat, forecast_samples,
and forecast_ci
}
\description{
Process state forecast
}
