% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_loc_forecast.R
\name{prop_from_timeseries}
\alias{prop_from_timeseries}
\title{Calculate proportion of disease ED visits by combining numerator
and denominator samples}
\usage{
prop_from_timeseries(
  e_denominator_samples,
  e_numerator_samples,
  required_columns
)
}
\description{
Calculate proportion of disease ED visits by combining numerator
and denominator samples
}
