% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_loc_forecast.R
\name{read_and_combine_data}
\alias{read_and_combine_data}
\title{Read in and combine daily and epiweekly training data
from a model run directory and calculate
proportion of ed visits}
\usage{
read_and_combine_data(model_run_dir)
}
\arguments{
\item{model_run_dir}{model run directory in which to look
for data.}
}
\description{
Read in and combine daily and epiweekly training data
from a model run directory and calculate
proportion of ed visits
}
