% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_hewr}
\alias{score_hewr}
\title{Default scoring function for \code{pyrenew-hew}.}
\usage{
score_hewr(
  scoreable_table,
  transform = scoringutils::log_shift,
  append_transformed = FALSE,
  offset = 1,
  ...
)
}
\arguments{
\item{scoreable_table}{Table to score, as the output
of \code{\link[=to_scoreable_table]{to_scoreable_table()}}.}

\item{transform}{transformation passed as the
\code{fun} argument to \code{\link[scoringutils:transform_forecasts]{scoringutils::transform_forecasts()}}.
Default \code{\link[scoringutils:log_shift]{scoringutils::log_shift()}}.}

\item{append_transformed}{When calling
\code{\link[scoringutils:transform_forecasts]{scoringutils::transform_forecasts()}}, append
the transformed scale forecasts to the base scale forecasts
or keep only the transformed scale forecasts? Passed as the
\code{append} argument to \code{\link[scoringutils:transform_forecasts]{scoringutils::transform_forecasts()}}.
Boolean, default \code{FALSE} (keep only transformed scale).}

\item{offset}{Offset for the transformation, passed to
\code{\link[scoringutils:transform_forecasts]{scoringutils::transform_forecasts()}}. Default 1.}

\item{...}{Other keyword arguments passed to
\code{\link[scoringutils:transform_forecasts]{scoringutils::transform_forecasts()}}.}
}
\value{
A table of scores, as the output of
\code{\link[scoringutils:score]{scoringutils::score()}}, filtered to include only the
transformed_scale.
}
\description{
Simple wrapper of \code{\link[scoringutils:transform_forecasts]{scoringutils::transform_forecasts()}}
and \code{\link[scoringutils:score]{scoringutils::score()}} with reasonable defaults for
\code{pyrenew-hew} production.
}
