% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_epiweekly_quantile_table.R
\name{to_epiweekly_quantile_table}
\alias{to_epiweekly_quantile_table}
\title{Create an epiweekly hubverse-format forecast quantile table
from a model batch directory containing forecasts
for multiple locations as daily MCMC draws.}
\usage{
to_epiweekly_quantile_table(
  model_batch_dir,
  exclude = NULL,
  epiweekly_other = FALSE,
  strict = FALSE
)
}
\arguments{
\item{model_batch_dir}{Model batch directory containing
the individual location forecast directories
("model run directories") to process. Name should be in the format
\verb{\{disease\}_r_\{reference_date\}_f_\{first_data_date\}_t_\{last_data_date\}}.}

\item{exclude}{Locations to exclude, if any, as a list of strings.
Default \code{NULL} (exclude nothing).}

\item{epiweekly_other}{Use an expressly epiweekly forecast
for non-target ED visits instead of a daily forecast aggregated
to epiweekly? Boolean, default \code{FALSE}.}

\item{strict}{Boolean. If \code{TRUE}, raise an error if no
valid draws are available to aggregate for any given location.
Otherwise return \code{NULL} for that location but continue with other.
locations. Passed as the \code{strict} argument to \code{\link[=to_epiweekly_quantiles]{to_epiweekly_quantiles()}}.
Default \code{FALSE}.}
}
\value{
The complete hubverse-format \code{\link[tibble:tibble]{tibble}}.
}
\description{
Create an epiweekly hubverse-format forecast quantile table
from a model batch directory containing forecasts
for multiple locations as daily MCMC draws.
}
