% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_epiweekly_quantile_table.R
\name{to_epiweekly_quantiles}
\alias{to_epiweekly_quantiles}
\title{Read in daily forecast draws from a model run directory
and output a set of epiweekly quantiles, as a
\code{\link[tibble:tibble]{tibbble}}.}
\usage{
to_epiweekly_quantiles(
  model_run_dir,
  report_date,
  max_lookback_days,
  disease_name = "prop_disease_ed_visits",
  draws_file_name = "epiweekly_samples",
  disease_model_name = "pyrenew_e",
  strict = FALSE
)
}
\arguments{
\item{model_run_dir}{Path to a directory containing
forecast draws to process, whose basename is the forecasted
location.}

\item{report_date}{Report date for which to generate epiweekly
quantiles.}

\item{max_lookback_days}{How many days before the report date
to look back when generating epiweekly quantiles (determines how
many negative epiweekly forecast horizons (i.e. nowcast/backcast)
quantiles will be generated.}

\item{disease_name}{Name of the disease quantity for which to
produced epiweekly quantiles. Default \code{"prop_disease_ed_visits"}.}

\item{draws_file_name}{Name of the parquet file containing
forecast draws. Default \code{"epiweekly_samples"}.}

\item{disease_model_name}{Name of the model for the target
disease. Default \code{"pyrenew_e"}.}

\item{strict}{Boolean. If \code{TRUE}, raise an error if no
valid draws are available to aggregate. Otherwise return
\code{NULL} in that case. Default \code{FALSE}.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} of quantiles, or \code{NULL}
if the draws file contains no vale
}
\description{
Read in daily forecast draws from a model run directory
and output a set of epiweekly quantiles, as a
\code{\link[tibble:tibble]{tibbble}}.
}
