% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{to_scoreable_table}
\alias{to_scoreable_table}
\title{Join hubverse forecast to observed data to
create a \code{scoringutils}-ready quantile forecast table.}
\usage{
to_scoreable_table(
  forecast,
  observed,
  horizons = c(0, 1),
  observed_value_column = "value",
  observed_location_column = "location",
  observed_date_column = "reference_date"
)
}
\arguments{
\item{forecast}{forecasts, as a hubverse-format
\code{\link[tibble:tibble]{tibble}} produced by
\code{\link[=to_epiweekly_quantile_table]{to_epiweekly_quantile_table()}}, with columns
\code{target_end_date}, \code{value}, and \code{horizon}}

\item{observed}{observations, as a \code{\link[tibble:tibble]{tibble}}.}

\item{horizons}{Horizons to score. Default \code{c(0, 1)}}

\item{observed_value_column}{Name of the column containing
observed values in the \code{observed} table, as a string.
Default \code{"value"}}

\item{observed_location_column}{Name of the column containing
location values in the \code{observed} table, as a string.
Default \code{"location"}}

\item{observed_date_column}{Name of the column containing
date values in the \code{observed} table, as a string.
Default \code{"reference_date"}}
}
\value{
A table for scoring, as the output of
\code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.
}
\description{
Expects forecast output created by \code{\link[=to_epiweekly_quantile_table]{to_epiweekly_quantile_table()}}
and an observed data table with location, date, and value columns.
The column names in the observed data table can be configured;
defaults are \code{"location"}, \code{"reference_date"}, and
\code{"value"}, respectively, to reflecting the format used
in pyrenew-hew pipeline scripts such as
\code{pipelines/create_observed_data_table.py}.
}
