% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_submission.R
\name{to_single_model_hubverse_table}
\alias{to_single_model_hubverse_table}
\title{Downsample a large hubverse table to a table
that contains only one particular model and only the
columns expected in a typical single model submission
to a hubverse-format Hub.}
\usage{
to_single_model_hubverse_table(df, model, targets = NULL)
}
\arguments{
\item{df}{A large format hubverse table, as the output
of \code{\link[=to_epiweekly_quantile_table]{to_epiweekly_quantile_table()}}}

\item{model}{Name of the model for which to get hubverse
submittable quantiles}

\item{targets}{Name(s) of the targets(s) for which to get
submittable quantiles. If \code{NULL}, get all included targets
from that model.}
}
\value{
The quantiles as a submittable hubverse table for the
given model.
}
\description{
Useful for creating Hub submissions.
}
