% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{to_tidy_draws_timeseries}
\alias{to_tidy_draws_timeseries}
\title{Combine a forecast in tidy draws based format
with observed values to create a synthetic set
of tidy posterior "samples".}
\usage{
to_tidy_draws_timeseries(
  tidy_forecast,
  observed,
  date_colname = "date",
  sample_id_colname = ".draw",
  value_colname = ".value",
  epiweekly = FALSE
)
}
\arguments{
\item{tidy_forecast}{Forecast in tidy format, with
a sample id column and a value column.}

\item{observed}{observed data to join with the forecast.}

\item{date_colname}{Name of the column in \code{tidy_forecast}
and \code{observed} that identifies dates. Default \code{"date"}.}

\item{sample_id_colname}{Name of the column in
\code{tidy_forecast} that uniquely identifies individual
posterior samples / draws. Default \code{".draw"}.}

\item{value_colname}{Name of the column in
\code{tidy_forecast} for the sampled values.
Default \code{".value"}.}

\item{epiweekly}{Is the timeseries epiweekly (as opposed
to daily)? Boolean, default \code{FALSE} (i.e. daily timeseries).}
}
\description{
Observed timepoints have the observed value as
the sampled value for all sample ids.
}
