% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{with_prop_disease_ed_visits}
\alias{with_prop_disease_ed_visits}
\title{Annotate a dataframe of ED visits data with the
proportion of visits due to a target disease.}
\usage{
with_prop_disease_ed_visits(df)
}
\arguments{
\item{df}{dataframe to annotate, with columns
\code{"Disease"} and \code{"Other"}.}
}
\value{
the dataframe with an additional column
\code{prop_disease_ed_visits}.
}
\description{
Annotate a dataframe of ED visits data with the
proportion of visits due to a target disease.
}
