% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast_figure.R
\name{make_forecast_figure}
\alias{make_forecast_figure}
\title{Make Forecast Figure}
\usage{
make_forecast_figure(
  dat,
  geo_value,
  disease,
  .variable,
  resolution,
  aggregated_numerator,
  aggregated_denominator,
  y_transform,
  ci,
  data_vintage_date,
  highlight_dates = NULL,
  highlight_labels = NULL,
  display_cutpoints = TRUE,
  max_lab_site_index = 5
)
}
\arguments{
\item{dat}{a data frame containing the data to be plotted}

\item{geo_value}{character matching the geo_value column in \code{dat} and \code{ci}}

\item{disease}{character matching the disease column in \code{dat} and \code{ci}}

\item{.variable}{character matching the .variable column in \code{dat} and \code{ci}}

\item{resolution}{character matching the resolution column in \code{dat} and \code{ci}}

\item{aggregated_numerator}{character matching the aggregated_numerator
column in \code{dat} and \code{ci}}

\item{aggregated_denominator}{character matching the aggregated_denominator
column in \code{dat} and \code{ci}}

\item{y_transform}{a character passed as the transform argument to
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}

\item{ci}{a data frame containing the credible intervals to be plotted}

\item{data_vintage_date}{date that the data was collected}

\item{highlight_dates}{a vector of dates to highlight on the plot}

\item{highlight_labels}{a vector of labels to display at the
\code{highlight_dates}}

\item{display_cutpoints}{a logical indicating whether to include cutpoints
relevant to the \code{target_variable}.}

\item{max_lab_site_index}{an integer indicating the maximum lab site index to
plot. Default is 5.}
}
\value{
a ggplot object
}
\description{
Make Forecast Figure
}
