% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast_figure.R
\name{make_forecast_figure}
\alias{make_forecast_figure}
\title{Make Forecast Figure}
\usage{
make_forecast_figure(
  target_disease,
  combined_dat,
  forecast_ci,
  disease_name = c("COVID-19", "Influenza"),
  data_vintage_date,
  y_transform = "identity"
)
}
\arguments{
\item{target_disease}{a disease matching the disease columns in combined_dat
and forecast_ci}

\item{combined_dat}{\code{combined_dat} from the result of process_state_forecast}

\item{forecast_ci}{\code{forecast_ci} from the result of process_state_forecast}

\item{disease_name}{"COVID-19" or "Influenza"}

\item{data_vintage_date}{date that the data was collected}

\item{y_transform}{a character passed as the transform argument to
ggplot2::scale_y_continuous()}
}
\value{
a ggplot object
}
\description{
Make Forecast Figure
}
