% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{process_state_forecast}
\alias{process_state_forecast}
\title{Process state forecast}
\usage{
process_state_forecast(
  model_run_dir,
  pyrenew_model_name,
  timeseries_model_name,
  save = TRUE
)
}
\arguments{
\item{model_run_dir}{Model run directory}

\item{pyrenew_model_name}{Name of directory containing pyrenew
model outputs}

\item{timeseries_model_name}{Name of directory containing timeseries
model outputs}

\item{save}{Logical indicating whether or not to save}
}
\value{
a list of 8 tibbles:
\code{combined_dat},
\code{epiweekly_combined_dat},
\code{forecast_samples},
\code{epiweekly_forecast_samples},
\code{forecast_with_epiweekly_other},
\code{forecast_ci},
\code{epiweekly_forecast_ci},
\code{forecast_with_epiweekly_other_ci}
}
\description{
Process state forecast
}
