% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{process_state_forecast}
\alias{process_state_forecast}
\title{Process state forecast}
\usage{
process_state_forecast(
  model_run_dir,
  n_forecast_days,
  pyrenew_model_name = NA,
  timeseries_model_name = NA,
  ci_widths = c(0.5, 0.8, 0.95),
  save = TRUE
)
}
\arguments{
\item{model_run_dir}{Model run directory}

\item{n_forecast_days}{An integer specifying the number of days to forecast.}

\item{pyrenew_model_name}{Name of directory containing pyrenew
model outputs}

\item{timeseries_model_name}{Name of directory containing timeseries
model outputs}

\item{ci_widths}{Vector of probabilities indicating one or more
central credible intervals to compute. Passed as the \code{.width}
argument to \code{\link[ggdist:point_interval]{ggdist::median_qi()}}. Default \code{c(0.5, 0.8, 0.95)}.}

\item{save}{Boolean indicating whether or not to save the output
to parquet files. Default \code{TRUE}.}
}
\value{
a list of 8 tibbles:
\code{daily_combined_training_eval_data},
\code{epiweekly_combined_training_eval_data},
\code{daily_samples},
\code{epiweekly_samples},
\code{epiweekly_with_epiweekly_other_samples},
\code{daily_ci},
\code{epiweekly_ci},
\code{epiweekly_with_epiweekly_other_ci}
}
\description{
Process state forecast
}
