% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_loc_forecast.R
\name{read_and_combine_data}
\alias{read_and_combine_data}
\title{Read in and combine training and evaluation
data from a model run directory.}
\usage{
read_and_combine_data(model_dir)
}
\arguments{
\item{model_dir}{model directory in which to look
for data.}
}
\description{
Read in and combine training and evaluation
data from a model run directory.
}
