% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{read_and_combine_data}
\alias{read_and_combine_data}
\title{Read in and combine training and evaluation
data from a model run directory.}
\usage{
read_and_combine_data(model_run_dir, epiweekly = FALSE)
}
\arguments{
\item{model_run_dir}{model run directoryh in which to look
for data.}

\item{epiweekly}{Get epiweekly data instead of daily data?
Boolean, default \code{FALSE}.}
}
\description{
Read in and combine training and evaluation
data from a model run directory.
}
