% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_epiweekly_quantile_table.R
\name{to_epiweekly_quantile_table}
\alias{to_epiweekly_quantile_table}
\title{Create an epiweekly hubverse-format forecast quantile table
from a model batch directory containing forecasts
for multiple locations as daily MCMC draws.}
\usage{
to_epiweekly_quantile_table(model_batch_dir)
}
\arguments{
\item{model_batch_dir}{Model batch directory containing
the individual location forecast directories
("model run directories") to process. Name should be in the format
\verb{\{disease\}_r_\{reference_date\}_f_\{first_data_date\}_t_\{last_data_date\}}.}
}
\value{
The complete hubverse-format \code{\link[tibble:tibble]{tibble}}.
}
\description{
Create an epiweekly hubverse-format forecast quantile table
from a model batch directory containing forecasts
for multiple locations as daily MCMC draws.
}
