% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_epiweekly_quantile_table.R
\name{to_epiweekly_quantiles}
\alias{to_epiweekly_quantiles}
\title{Read in daily forecast draws from a model run directory
and output a set of epiweekly quantiles, as a
\code{\link[tibble:tibble]{tibbble}}.}
\usage{
to_epiweekly_quantiles(model_run_dir, report_date, max_lookback_days)
}
\arguments{
\item{model_run_dir}{Path to a directory containing
forecast draws to process, whose basename is the forecasted
location.}

\item{report_date}{Report date for which to generate epiweekly quantiles.}

\item{max_lookback_days}{How many days before the report date
to look back when generating epiweekly quantiles (determines how
many negative epiweekly forecast horizons (i.e. nowcast/backcast)
quantiles will be generated.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} of quantiles.
}
\description{
Read in daily forecast draws from a model run directory
and output a set of epiweekly quantiles, as a
\code{\link[tibble:tibble]{tibbble}}.
}
