% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_forecast.R
\name{with_prop_disease_ed_visits}
\alias{with_prop_disease_ed_visits}
\title{Compute the proportion of ED visits due to
the target disease.}
\usage{
with_prop_disease_ed_visits(df)
}
\arguments{
\item{df}{dataframe to annotate, with coulumns
\code{"Disease"} and \code{"Other"}.}
}
\value{
the dataframe with an additional column
\code{prop_disease_ed_visits}.
}
\description{
Compute the proportion of ED visits due to
the target disease.
}
